requirements.txt
torch>=2.0.0
transformers>=4.25.0
Pillow>=9.0.0
gradio>=4.0.0
tqdm>=4.64.0
numpy>=1.24.0
matplotlib>=3.6.0
datasets>=2.14.0
accelerate>=0.21.0
safetensors>=0.3.0 